feat: str as indented list
In trees.py, three methods are modified in PeakTree and inherited by HyperPeakTree:
subtree() yields tree nodes in a different depth-first order.
This new order is used in the dict returned by levels(), and levels() is faster.
This levels dict enables str() to print the tree in the indented list notation, where the level is the indentation.